EDIT: + API:  - count rating by rate: change Rate from double to integer       + Response: RateWithCountRatingResponse: change from Double rate to Integer rate       + Service & Repository: change argument from Double rate to Integer rate       + SupportFunctions: getAllRating

+++ src/main/java/swp490/spa/dto/responses/RateWithCountRatingResponse.java
@@ -10,7 +10,7 @@ import lombok.*;
 @Setter
 public class RateWithCountRatingResponse {
     @JsonProperty("rate")
-    private Double rate;
+    private Integer rate;
     @JsonProperty("count_rating")
     private Integer countRating;
 }
